+++ samples/WpfApp1/MainWindow.xaml.cs
@@ -298,7 +298,7 @@ public partial class MainWindow : Window
         };
         if (ofd.ShowDialog(this) == true)
         {
-            var ofd2 = new SaveFileDialog()
+            var ofd2 = new OpenFileDialog()
             {
                 Filter = "Word OpenXML document|*.docx",
                 FileName = Path.GetFileNameWithoutExtension(ofd.FileName) + ".docx"