+++ src/DocSharp.Markdown/Docx/Extensions/TableRenderer.cs
@@ -11,47 +11,24 @@ public class TableRenderer : DocxObjectRenderer<Markdig.Extensions.Tables.Table>
         renderer.ForceCloseParagraph();
 
         var table = new Table();
-        var tableProperties = new TableProperties(
-            new TableBorders(
-                new TopBorder { Val = new EnumValue<BorderValues>(BorderValues.Single), Size = 5 },
-                new BottomBorder { Val = new EnumValue<BorderValues>(BorderValues.Single), Size = 5 },
-                new LeftBorder { Val = new EnumValue<BorderValues>(BorderValues.Single), Size = 5 },
-                new RightBorder { Val = new EnumValue<BorderValues>(BorderValues.Single), Size = 5 },
-                new InsideHorizontalBorder { Val = new EnumValue<BorderValues>(BorderValues.Single), Size = 5 },
-                new InsideVerticalBorder { Val = new EnumValue<BorderValues>(BorderValues.Single), Size = 5 }
-            )
-        );
+        var tableProperties = new TableProperties()
+        {
+            TableStyle = new TableStyle() { Val = "MDTable" },
+        };
         table.Append(tableProperties);
         renderer.Cursor.Write(table);
 
-        bool firstRow = true;
         foreach (var row in obj.OfType<Markdig.Extensions.Tables.TableRow>())
         {
             var tableRow = new TableRow();
-            if (firstRow)
-            {
-                tableRow.AddChild(new TableRowProperties(new TableHeader()));
-            }
             table.Append(tableRow);
             foreach (var cell in row.OfType<Markdig.Extensions.Tables.TableCell>())
             {
                 var tableCell = new TableCell();
-                if (firstRow)
-                {
-                    tableCell.Append(new TableCellProperties()
-                    {
-                        Shading = new Shading() { 
-                            Color = "auto", 
-                            Fill = "D9D9D9", 
-                            Val = ShadingPatternValues.Clear,                           
-                        }
-                    });
-                }
                 tableRow.Append(tableCell);
                 renderer.Cursor.GoInto(tableCell);
                 renderer.WriteChildren(cell);
             }
-            firstRow = false;
         }
         renderer.Cursor.SetAfter(table);
     }